Various improvements to the input system
Player input is sent to the server inmediately (which is important) and the position is interpolated afterwards. Inputs are also more precise, causing virtually cero differences in position between client and server.

Input ids use 32 bits now.

Server now clamps inputRate and snapshotRate values instead of not accepting them at all if they we're invalid.

+++ include/game_server.hpp
@@ -34,7 +34,7 @@ public:
 
         u32 snapshotId = 0;
         sf::Time snapshotRate;
-        u16 latestInputId = 0;
+        u32 latestInputId = 0;
         sf::Time inputRate;
         u8 inputsSent = 0; //this update
